+++ example/lib/pages/input_form_field.dart
@@ -72,7 +72,7 @@ class _InputFormFieldPageState extends State<InputFormFieldPage> {
               children: [
                 ShadInputFormField(
                   id: 'username',
-                  prefix: const Icon(LucideIcons.user),
+                  leading: const Icon(LucideIcons.user),
                   label: const Text('Username'),
                   placeholder: const Text('Enter your username'),
                   description: const Text('This is your public display name.'),